% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connection}
\usage{
connect(url, version, ...)
}
\arguments{
\item{url}{Base url, e.g., \code{http://localhost:8080}}

\item{version}{API version, e.g, \code{v1}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Connection
}
\examples{
library("httr")
(conn <- connect())
conn$url
conn$version
conn$content_type
conn$status()
conn$status(config = verbose())
conn$routes()
conn$routes(config = verbose())
}

