% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonAPIModel.R
\name{japi_model}
\alias{japi_model}
\title{JSON API Model and Serializer}
\usage{
japi_model(...)
}
\arguments{
\item{...}{named parameters}
}
\description{
JSON API Model and Serializer
}
\examples{
mod <- japi_model("id", "name", "year", "actor_ids", 
  "owner_id", "movie_type_id")
mod

# assign values
mod$id = 232
mod$name = 'test movie'
mod$actor_ids = 1:3
mod$movie_type_id = 1
mod

# put into json api serializer
x <- japi_serializer$new(
  model = mod,
  type  = "movie",
  attributes = c('name', 'year'),
  has_many = "actors",
  belongs_to = "movie_type"
 )
x$model
x$attributes
x$json_api_model
x$serialized_json()
x$serialized_json(pretty = TRUE)
}
