% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonAPIModel.R
\docType{data}
\name{japi_serializer}
\alias{japi_serializer}
\title{JSON API Serializer}
\arguments{
\item{model}{an object of class `japi_model`}

\item{type}{the type}

\item{attributes}{attributes}

\item{has_many}{variables that are many}

\item{belongs_to}{belongs to type}
}
\description{
JSON API Serializer
}
\details{
see \code{\link{japi_model}} for examples
}
\keyword{datasets}
