% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jsonapi-package.R
\docType{package}
\name{jsonapi-package}
\alias{jsonapi}
\alias{jsonapi-package}
\title{jsonapi client}
\description{
jsonapi client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

