% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjsonapi-package.R
\docType{package}
\name{rjsonapi-package}
\alias{rjsonapi}
\alias{rjsonapi-package}
\title{rjsonapi client}
\description{
rjsonapi client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

