% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connection}
\usage{
connect(url, version, endpt, query, ...)
}
\arguments{
\item{url}{(character)Base url, e.g., \code{http://localhost:8080}}

\item{version}{(character) API version, e.g, \code{v1}}

\item{endpt}{(character) The route, or endpoint}

\item{query}{(list) A list of query parameters}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Connection
}
\examples{
\dontrun{
library("httr")
(conn <- connect("http://localhost:8088"))
conn$url
conn$version
conn$content_type
conn$status()
conn$status(config = verbose())
conn$routes()
conn$routes(config = verbose())

# get data from speicific routes
conn$route("authors")
conn$route("chapters")
conn$route("authors/1")
conn$route("authors/1/books")
conn$route("chapters/5")
conn$route("chapters/5/book")
conn$route("chapters/5/relationships/book")

## include
conn$route("authors/1", include = "books")
conn$route("authors/1", include = "photos")
conn$route("authors/1", include = "photos.title")

## errors
### route doesn't exist
conn$route("foobar")

### document doesn't exist
conn$route("authors/56")


(conn <- connect("https://api.labs.datacite.org"))
conn$url
conn$version
conn$content_type
conn$status()
conn$status(config = verbose())
conn$route("members")
conn$route("datacentres")
conn$route("works")
## query parameters
aa <- conn$route("works", query = list(q = "publicationYear:[1980 TO 1982]"))
unlist(aa$data$attributes$issued$`date-parts`)
}
}

