% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JsonAPIModel.R
\docType{data}
\name{japi_model}
\alias{japi_model}
\alias{japi_serializer}
\title{JSON API Model and Serializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
japi_model(...)

japi_serializer
}
\arguments{
\item{...}{named parameters}

\item{model}{an object of class `japi_model`}

\item{type}{the type}

\item{attributes}{attributes}

\item{has_many}{variables that are many}

\item{belongs_to}{belongs to type}
}
\description{
JSON API Model and Serializer
}
\examples{
mod <- japi_model("id", "name", "year", "actor_ids", 
  "owner_id", "movie_type_id")
mod

# assign values
mod$id = 232
mod$name = 'test movie'
mod$actor_ids = 1:3
mod$movie_type_id = 1
mod

# put into json api serializer
x <- JApiSerializer$new(
  model = mod,
  type  = "movie",
  attributes = c('name', 'year'),
  has_many = "actors",
  belongs_to = "movie_type"
 )
x$model
x$attributes
x$json_api_model
x$serialized_json()
x$serialized_json(pretty = TRUE)
}
\keyword{datasets}
