% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjsonapi-package.R
\docType{package}
\name{rjsonapi-package}
\alias{rjsonapi-package}
\alias{rjsonapi}
\title{rjsonapi}
\description{
Consumer for APIs that follow the JSON API specification
}
\section{JSON API}{

JSON API is a specification for building APIs (Application Programming
Interfaces) in JSON (Javascript Object Notation).

The JSON API website is at https://jsonapi.org/ . The specification
itself is at https://jsonapi.org/format/, with implementations at
https://jsonapi.org/implementations/
}

\section{Examples}{

There are very few examples to see in the real world. One example
comes from CodeClimate. Their API docs are at
https://developer.codeclimate.com/#overview
}

\author{
Scott Chamberlain
}
\keyword{package}
