% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjsonapi-package.R
\docType{package}
\name{rjsonapi-package}
\alias{rjsonapi-package}
\alias{rjsonapi}
\title{rjsonapi client}
\description{
rjsonapi client
}
\section{JSON API}{

JSON API is a specification for building APIs (Application Programming
Interfaces) in JSON (Javascript Object Notation).

The JSON API website is at \url{http://jsonapi.org/}. The specification
itself is at \url{http://jsonapi.org/format/}, with implementations at
\url{http://jsonapi.org/implementations/}.
}

\section{Examples}{

There are very few examples to see in the real world. One example
comes from CodeClimate. Their API docs are at
https://docs.codeclimate.com/docs/api
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
